Improved display

+++ src/Pret.java
@@ -17,6 +17,7 @@ public class Pret {
     private Connection con;
     Scanner sc, sc1;
     DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+    String format = "| %-5s | %-15s | %-15s | %-27s | %-27s |\n";
 
     public Pret() {
         try {
@@ -29,10 +30,8 @@ public class Pret {
     public void printPret() {
         sc = new Scanner(System.in);
         System.out.println("\nListe des prêts :\n");
-    
-        String format = "| %-5s | %-10s | %-10s | %-12s | %-12s |\n";
-        System.out.printf(format, "ID", "Etudiant", "Ouvrage", "Date prêt", "Date remise");
-        System.out.println("---------------------------------------------------------------");
+        System.out.printf(format, "ID", "Code Etudiant", "Code Ouvrage", "Date prêt", "Date remise");
+        System.out.println("---------------------------------------------------------------------------------------------------------");
     
         try {
             Statement st = con.createStatement();
@@ -63,10 +62,10 @@ public class Pret {
         numeroOuvrage = sc.nextInt();
         // System.out.println("Enter the loan date (YYYY-MM-DD): ");
         datePret = LocalDate.now(); // Use current date as loan date
-        // datePret = sc.next(); // Uncomment this line if you want to input the loan
-        // date manually
+        // datePret = sc.next(); // Uncomment this line if you want to input the loan return date manually
         System.out.println("Entrez le nombre de jours avant la date de retour: ");
         dateRemise = LocalDate.now().plusDays(sc.nextInt());
+        sc.nextLine(); 
         // String formattedDatePret = datePret.format(formatter);
         // String formattedDateRemise = dateRemise.format(formatter);
         String request = "INSERT INTO Loan (numeroEtudiant, numeroOuvrage, datePret, dateRemise) VALUES ("
@@ -84,9 +83,23 @@ public class Pret {
     }
 
     public void deletePret() {
+        int found;
         sc = new Scanner(System.in);
-        System.out.println("Enter the loan number to delete: ");
-        numeroPret = sc.nextInt();
+        found = searchPret();
+        if (found == -1) {
+            return;
+        }
+        // System.out.println("Enter the loan number to delete: ");
+        // numeroPret = sc.nextInt();
+        System.out.println("Are you sure you want to delete the loan with code " + numeroPret + "? (y/n)");
+        String confirmation = sc.nextLine();
+        // sc.nextLine();
+        if (!confirmation.equals("y") && !confirmation.equals("Y")) {
+            System.out.println("Deletion cancelled.");
+            System.out.println("Press Enter to continue");
+            sc.nextLine();
+            return;
+        }
         String request = "DELETE FROM Loan WHERE numeroPret = " + numeroPret;
         try {
             Statement st = con.createStatement();
@@ -102,16 +115,25 @@ public class Pret {
     int searchPret() {
         boolean found = false;
         sc = new Scanner(System.in);
-        System.out.println("Enter the loan number: ");
+        System.out.println("Enter the loan number (ID): ");
         numeroPret = sc.nextInt();
+        sc.nextLine();
         String request = "SELECT * FROM Loan WHERE numeroPret = " + numeroPret;
+        System.out.println("\nLe prêt :\n");
+        System.out.printf(format, "ID", "Code Etudiant", "Code Ouvrage", "Date prêt", "Date remise");
+        System.out.println("---------------------------------------------------------------------------------------------------------");
         try {
             Statement st = con.createStatement();
             ResultSet rs = st.executeQuery(request);
             if (rs.next()) {
                 found = true;
-                System.out.println(rs.getInt(1) + " - " + rs.getInt(2) + " - " + rs.getInt(3) + " - " + rs.getString(4)
-                        + " - " + rs.getString(5));
+                System.out.printf(format,
+                    rs.getInt(1),
+                    rs.getInt(2),
+                    rs.getInt(3),
+                    rs.getString(4),
+                    rs.getString(5)
+                );
             } else {
                 System.out.println("Loan not found");
             }
@@ -130,6 +152,7 @@ public class Pret {
         sc = new Scanner(System.in);
         System.out.println("Enter the loan number to update: ");
         numeroPret = sc.nextInt();
+        sc.nextLine();
         String request = "SELECT * FROM Loan WHERE numeroPret = " + numeroPret;
         try {
             Statement st = con.createStatement();
@@ -143,6 +166,7 @@ public class Pret {
                 datePret = LocalDate.now(); // Use current date as loan date
                 System.out.println("Entrez le nouveau nombre de jours avant la date de retour: ");
                 dateRemise = LocalDate.now().plusDays(sc.nextInt()); // Use current date plus days as return date
+                sc.nextLine();
                 request = "UPDATE Loan SET numeroEtudiant = ?, numeroOuvrage = ?, datePret = ?, dateRemise = ? WHERE numeroPret = ?";
                 PreparedStatement ps = con.prepareStatement(request);
                 ps.setInt(1, numeroEtudiant);
@@ -163,32 +187,20 @@ public class Pret {
         sc.nextLine();
     }
 
-    void getAll() {
-        try {ResultSet rs = con.getMetaData().getTables(null, null, null, new String[] { "TABLE" });
-        while (rs.next()) {
-            System.out.println("TABLE: " + rs.getString("TABLE_NAME"));
-        }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-
-    }
-
     public void menu() {
         Library library = new Library();
         String choice;
         sc = new Scanner(System.in);
         do {
-            // library.clearConsole();
+            library.clearConsole();
             System.out.println("Module Pret");
             System.out.println(
-                    "\n[1]. Add Loan\n[2]. Delete Loan\n[3]. Update Loan\n[4]. Search Loan\n[5]. List Loans\n[6]. Retour au menu principal\n");
+                    "\n[1]. Add Loan\n[2]. Delete Loan or Return book\n[3]. Update Loan\n[4]. Search Loan\n[5]. List Loans\n[6]. Retour au menu principal\n");
             System.out.println("Enter your choice: ");
             choice = sc.nextLine();
             switch (choice) {
                 case "1":
                     addPret();
-                    // getAll();
                     break;
                 case "2":
                     deletePret();
